DEV: Convert `watched-word-test` modal to component-based API
