+++ app/assets/javascripts/admin/addon/components/modal/watched-word-test.hbs
@@ -1,0 +1,45 @@
+<DModal
+  @title={{i18n
+    "admin.watched_words.test.modal_title"
+    action=@model.watchedWord.name
+  }}
+  @closeModal={{@closeModal}}
+  class="watched-words-test-modal"
+>
+  <:body>
+    <p>{{i18n "admin.watched_words.test.description"}}</p>
+    <Textarea name="test_value" @value={{this.value}} autofocus="autofocus" />
+    {{#if this.matches}}
+      <p>
+        {{i18n "admin.watched_words.test.found_matches"}}
+        <ul>
+          {{#if (or this.isReplace this.isLink)}}
+            {{#each this.matches as |match|}}
+              <li>
+                <span class="match">{{match.match}}</span>
+                &rarr;
+                <span class="replacement">{{match.replacement}}</span>
+              </li>
+            {{/each}}
+          {{else if this.isTag}}
+            {{#each this.matches as |match|}}
+              <li>
+                <span class="match">{{match.match}}</span>
+                &rarr;
+                {{#each match.tags as |tag|}}
+                  <span class="tag">{{tag}}</span>
+                {{/each}}
+              </li>
+            {{/each}}
+          {{else}}
+            {{#each this.matches as |match|}}
+              <li>{{match}}</li>
+            {{/each}}
+          {{/if}}
+        </ul>
+      </p>
+    {{else}}
+      <p>{{i18n "admin.watched_words.test.no_matches"}}</p>
+    {{/if}}
+  </:body>
+</DModal>
