--- app/assets/javascripts/admin/addon/templates/modal/admin-watched-word-test.hbs
@@ -1,42 +1,0 @@
-<DModalBody
-  @rawTitle={{i18n
-    "admin.watched_words.test.modal_title"
-    action=this.model.name
-  }}
-  @class="watched-words-test-modal"
->
-  <p>{{i18n "admin.watched_words.test.description"}}</p>
-  <Textarea name="test_value" @value={{this.value}} autofocus="autofocus" />
-  {{#if this.matches}}
-    <p>
-      {{i18n "admin.watched_words.test.found_matches"}}
-      <ul>
-        {{#if (or this.isReplace this.isLink)}}
-          {{#each this.matches as |match|}}
-            <li>
-              <span class="match">{{match.match}}</span>
-              &rarr;
-              <span class="replacement">{{match.replacement}}</span>
-            </li>
-          {{/each}}
-        {{else if this.isTag}}
-          {{#each this.matches as |match|}}
-            <li>
-              <span class="match">{{match.match}}</span>
-              &rarr;
-              {{#each match.tags as |tag|}}
-                <span class="tag">{{tag}}</span>
-              {{/each}}
-            </li>
-          {{/each}}
-        {{else}}
-          {{#each this.matches as |match|}}
-            <li>{{match}}</li>
-          {{/each}}
-        {{/if}}
-      </ul>
-    </p>
-  {{else}}
-    <p>{{i18n "admin.watched_words.test.no_matches"}}</p>
-  {{/if}}
-</DModalBody>
