+++ app/assets/javascripts/discourse/app/services/modal.js
@@ -48,7 +48,6 @@ const KNOWN_LEGACY_MODALS = [
   "tag-upload",
   "topic-summary",
   "user-status",
-  "admin-watched-word-test",
   "admin-install-theme",
   "admin-penalize-user",
   "admin-theme-change",
